WIP on saveAndCreateJob in build-answer-set controller
- Moves steps to save stage groups and dds files out of controller
- Creates (and saves) a job-stage-group at time of creating the job-answer-set. May not be necessary, may have been a result of the next item.
- Adds missing provideInputFiles parameter to FileGroupList (this was missing and it caused the parent component to not be able to call this action)
- Saves inputFile on provide, but save doesn't work correctly the first time - the promise for the credential is not fulfilled, so the JSON payload has null for this field immediately. For now, clicking provide twice works.
- Fixes bug in the redirect from /jobs/new -> /jobs/new/select-workflow. This needed to happen in /jobs/new/index so it didn't stomp on all other child routes of jobs/new
