+++ app/routes/jobs/new.js
@@ -1,7 +1,4 @@
 import Ember from 'ember';
 
 export default Ember.Route.extend({
-  beforeModel() {
-    this.transitionTo('jobs.new.select-workflow');
-  }
 });
