+++ app/routes/jobs/new/index.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  beforeModel() {
+    this.transitionTo('jobs.new.select-workflow');
+  }
+});
