+++ app/templates/jobs/new/index.hbs
@@ -1,0 +1,1 @@
+{{outlet}}
