+++ tests/unit/routes/jobs/new-test.js
@@ -7,12 +7,3 @@ test('it exists', function(assert) {
   let route = this.subject();
   assert.ok(route);
 });
-
-test('it transitions to workflow selection', function(assert) {
-  let route = this.subject({
-    transitionTo(routeName) {
-      assert.equal('jobs.new.select-workflow', routeName);
-    }
-  });
-  route.beforeModel();
-});
